requirements.txt
tqdm
svg.path